Allow an options hash to be passed in to Upcoming Invoice API
The Upcoming Invoice API now supports a ton of cool optional params
to help calculate prorations.

Since the customer ID is still required, and the subscription id is
optional, the second parameter to the retrieveUpcoming endpoint
can be an options object or a string (string must be the subscription
id for legacy compatibility).

Fixes #176